Reorganize imports and refactor hf and openai modules for better structure
