+++ pyproject.toml
@@ -68,7 +68,7 @@ extend-safe-fixes = [
   "Q",    # flake8-quotes
 ]
 extend-fixable = ["B", "D"]
-ignore = ["T201", "T203"] # print, pprint statements OK
+ignore = ["T201", "T203", "F401"] # print, pprint statements OK, unused imports OK
 isort.case-sensitive = true
 flake8-type-checking.quote-annotations = true
 pydocstyle.convention = "google"
