+++ src/huggingface/__init__.py
@@ -1,0 +1,1 @@
+from .hf_utils import login_hf_hub
