+++ src/huggingface/hf_inference.py
@@ -1,0 +1,1 @@
+"""Module for running inference on a Hugging Face model."""
